Refactor ETA calculation
